requirements.txt
unstructured[local-inference]
PyPDF2
langchain-huggingface
langchain-community
chromadb
langchain-google-genai
python-dotenv
tiktoken